LayoutTests/imported/w3c/web-platform-tests/html/dom/elements/global-attributes/the-anchor-attribute-xml.tentative-expected.txt


FAIL the-anchor-attribute-xml assert_equals: Setting the anchor attribute in XML should work. expected (object) Element node <div id="target">target</div> but got (undefined) undefined
